Include py-vox-io in package to avoid installing
The MIT license in the pyvox module allows including it in other
projects, so there is no need to install it from PyPI. Numpy does not
need to be installed either, because it is always included in Blender,
since some built-in addons use it.

This change eliminates the need to launch Blender with elevated
privileges in order to install dependencies, as well as the risk of
breaking the addon or Blender itself by installing the wrong version
of a package.